implement test algorithm

+++ HW#3/paging-multilevel-translate.py
@@ -42,7 +42,7 @@ class OS:
         # os tracks
         self.usedPages      = []
         self.usedPagesCount = 0
-        self.maxPageCount   = self.physMem / self.pageSize
+        self.maxPageCount   = int(self.physMem / self.pageSize)
 
         # no pages used (yet)
         for i in range(0, self.maxPageCount):
@@ -187,7 +187,7 @@ class OS:
         print('')
 
     def memoryDump(self):
-        for i in range(0, self.physMem / self.pageSize):
+        for i in range(0, int(self.physMem / self.pageSize)):
             print('page %3d:' %  i, end=' ') 
             for j in range(0, self.pageSize):
                 print('%02x' % self.memory[(i * self.pageSize) + j], end=' ')
